+++ component/loki/process/internal/stages/metric.go
@@ -69,7 +69,10 @@ func newMetricStage(logger log.Logger, config MetricsConfig, registry prometheus
 			if err != nil {
 				return nil, err
 			}
-			registry.MustRegister(collector)
+			collector, err = registerOrReuse(registry, collector)
+			if err != nil {
+				return nil, err
+			}
 			metrics[cfg.Counter.Name] = cfgCollector{cfg: cfg, collector: collector}
 		case cfg.Gauge != nil:
 			customPrefix := ""
@@ -82,7 +85,10 @@ func newMetricStage(logger log.Logger, config MetricsConfig, registry prometheus
 			if err != nil {
 				return nil, err
 			}
-			registry.MustRegister(collector)
+			collector, err = registerOrReuse(registry, collector)
+			if err != nil {
+				return nil, err
+			}
 			metrics[cfg.Gauge.Name] = cfgCollector{cfg: cfg, collector: collector}
 		case cfg.Histogram != nil:
 			customPrefix := ""
@@ -95,7 +101,10 @@ func newMetricStage(logger log.Logger, config MetricsConfig, registry prometheus
 			if err != nil {
 				return nil, err
 			}
-			registry.MustRegister(collector)
+			collector, err = registerOrReuse(registry, collector)
+			if err != nil {
+				return nil, err
+			}
 			metrics[cfg.Histogram.Name] = cfgCollector{cfg: cfg, collector: collector}
 		default:
 			return nil, fmt.Errorf("undefined stage type in '%v', exiting", cfg)
@@ -106,6 +115,17 @@ func newMetricStage(logger log.Logger, config MetricsConfig, registry prometheus
 		cfg:     config,
 		metrics: metrics,
 	}), nil
+}
+
+func registerOrReuse(registerer prometheus.Registerer, c prometheus.Collector) (prometheus.Collector, error) {
+	err := registerer.Register(c)
+	if err == nil {
+		return c, nil
+	}
+	if previous, ok := err.(prometheus.AlreadyRegisteredError); ok {
+		return previous.ExistingCollector, nil
+	}
+	return nil, err
 }
 
 // metricStage creates and updates prometheus metrics based on extracted pipeline data