+++ CHANGELOG.md
@@ -108,6 +108,8 @@ Main (unreleased)
 
 - Fix issue where the UI could not navigate to components loaded by modules. (@rfratto)
 
+- Fix issue where updating some modules' config (e.g. `loki.process`) could lead to a panic. (@thampiotr) 
+
 ### Other changes
 
 - Add metrics when clustering mode is enabled. (@rfratto)